+++ src/main/kotlin/mc/fuckoka/economyapi/domain/model/WalletID.kt
@@ -1,6 +1,6 @@
 package mc.fuckoka.economyapi.domain.model
 
-data class WalletID(val value: Long) {
+data class WalletID(val value: Int) {
     init {
         require(value >= 0)
     }
